module/StableDiffsuion/requirements.txt
absl-py==1.4.0
aiohttp==3.8.3
aiosignal==1.2.0
antlr4-python3-runtime==4.8
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
cachetools==4.2.4
certifi==2022.12.7
charset-normalizer==2.0.12
click==8.0.4
clip @ git+https://github.com/openai/CLIP.git@3702849800aa56e2223035bccd1c6ef91c704ca8
cycler==0.11.0
dataclasses==0.6
decorator==4.4.2
einops==0.3.0
filelock==3.4.1
flatbuffers==23.1.21
frozenlist==1.2.0
fsspec==2022.1.0
ftfy==6.0.3
future==0.18.3
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.48.2
huggingface-hub==0.4.0
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
imageio-ffmpeg==0.4.8
importlib-metadata==4.8.3
importlib-resources==5.4.0
jedi==0.18.2
joblib==1.1.1
kiwisolver==1.3.1
kornia==0.6.8
Markdown==3.3.7
matplotlib==3.3.4
moviepy==1.0.3
multidict==5.2.0
networkx==2.5.1
numpy==1.19.5
oauthlib==3.2.2
omegaconf==2.1.1
onnxruntime==1.12.0
opencv-python==4.8.0.74
packaging==21.3
pandas==1.1.5
parso==0.8.3
Pillow==8.4.0
proglog==0.1.10
protobuf==3.19.0
pudb==2022.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyDeprecate==0.3.1
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytorch-lightning==1.4.2
pytz==2022.7.1
PyWavelets==1.1.1
PyYAML==6.0
regex==2022.10.31
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
sacremoses==0.0.53
scikit-image==0.17.2
scipy==1.5.4
six==1.16.0
taming-transformers==0.0.1
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2020.9.3
tokenizers==0.12.1
torch==1.8.1+cu101
torchaudio==0.8.1
torchmetrics==0.6.0
torchvision==0.9.1+cu101
tqdm==4.64.1
transformers==4.18.0
typing_extensions==4.1.1
urllib3==1.26.14
urwid==2.1.2
urwid-readline==0.13
wcwidth==0.2.6
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.6.0